+++ src/components/character/newCharacter.dto.ts
@@ -3,4 +3,5 @@ import { Gender } from "models/gender.model";
 export class NewCharacterDto {
   nick: string;
   gender: Gender;
+  sprite?: string;
 }